[Module] - Handle file reading and error handling
