+++ main.py
@@ -1,6 +1,7 @@
 import csv
 import time as datetime
 import json
+from utils import read_file, test_valid_file_name_and_extension
 
 class BadRequest(Exception):
 
@@ -11,19 +12,6 @@ class BadRequest(Exception):
 
     def __str__(self):
         return self.message
-
-def read_file():
-    filename = input("Enter filename:")
-    if not filename.strip():
-        raise BadRequest("Cannot be blank.")
-    file = open(filename)
-    return file
-
-def test_valid_file_name_and_extension(filename, extension):
-  if filename.split(".")[1:] in [[extension]] and filename[0].isalpha() and sum(c.isdigit() for c in filename) < 4:
-    return True
-  else:
-    raise BadRequest("Invalid File")
 
 def buy_or_sell(flag):
     if (flag == "b"):
